+++ cfripper/rules/__init__.py
@@ -12,25 +12,25 @@ under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-from .CrossAccountTrustRule import CrossAccountTrustRule
-from .EBSVolumeHasSSERule import EBSVolumeHasSSERule
-from .FullWildcardPrincipalRule import FullWildcardPrincipalRule
-from .HardcodedRDSPasswordRule import HardcodedRDSPasswordRule
-from .IAMRolesOverprivilegedRule import IAMRolesOverprivilegedRule
-from .KMSKeyWildcardPrincipal import KMSKeyWildcardPrincipal
-from .ManagedPolicyOnUserRule import ManagedPolicyOnUserRule
-from .PartialWildcardPrincipal import PartialWildcardPrincipalRule
-from .PolicyOnUserRule import PolicyOnUserRule
-from .PrivilegeEscalationRule import PrivilegeEscalationRule
-from .S3BucketPolicyPrincipalRule import S3BucketPolicyPrincipalRule
-from .S3BucketPublicReadAclAndListStatementRule import S3BucketPublicReadAclAndListStatementRule
-from .S3BucketPublicReadWriteAclRule import S3BucketPublicReadWriteAclRule
-from .S3CrossAccountTrustRule import S3CrossAccountTrustRule
-from .SecurityGroupIngressOpenToWorld import SecurityGroupIngressOpenToWorld
-from .SecurityGroupOpenToWorldRule import SecurityGroupOpenToWorldRule
-from .SNSTopicPolicyNotPrincipalRule import SNSTopicPolicyNotPrincipalRule
-from .SQSQueuePolicyNotPrincipalRule import SQSQueuePolicyNotPrincipalRule
-from .SQSQueuePolicyPublicRule import SQSQueuePolicyPublicRule
+from cfripper.rules.cross_account_trust import S3CrossAccountTrustRule
+from cfripper.rules.s3_public_access import S3BucketPublicReadWriteAclRule
+from cfripper.rules.security_group import SecurityGroupOpenToWorldRule
+from cfripper.rules.sqs_queue_policy import SQSQueuePolicyPublicRule
+
+from .cross_account_trust import CrossAccountTrustRule
+from .ebs_volume_has_sse import EBSVolumeHasSSERule
+from .hardcoded_RDS_password import HardcodedRDSPasswordRule
+from .iam_roles import IAMRolesOverprivilegedRule
+from .kms_key_wildcard_principal import KMSKeyWildcardPrincipal
+from .managed_policy_on_user import ManagedPolicyOnUserRule
+from .policy_on_user import PolicyOnUserRule
+from .privilege_escalation import PrivilegeEscalationRule
+from .s3_bucked_policy import S3BucketPolicyPrincipalRule
+from .s3_public_access import S3BucketPublicReadAclAndListStatementRule
+from .security_group import SecurityGroupIngressOpenToWorld
+from .sns_topic_policy_not_principal import SNSTopicPolicyNotPrincipalRule
+from .sqs_queue_policy import SQSQueuePolicyNotPrincipalRule
+from .wildcard_principals import FullWildcardPrincipalRule, PartialWildcardPrincipalRule
 
 DEFAULT_RULES = {
     "IAMRolesOverprivilegedRule": IAMRolesOverprivilegedRule,